Перенесет функцию sortOffers в utils.js
