+++ src/utils.js
@@ -1,3 +1,5 @@
+import {SortTypes} from "./const.js";
+
 const getRatingWidth = (rating) => {
   return Math.round(rating) * 20;
 };
@@ -6,4 +8,25 @@ const extend = (a, b) => {
   return Object.assign({}, a, b);
 };
 
-export {getRatingWidth, extend};
+const sortOffers = (offers, sortType) => {
+  let sortedOffers = [];
+
+  switch (sortType) {
+    case SortTypes.POPULAR:
+      sortedOffers = offers;
+      break;
+    case SortTypes.PRICE_LOW_HIGH:
+      sortedOffers = offers.slice().sort((a, b) => a.price - b.price);
+      break;
+    case SortTypes.PRICE_HIGH_LOW:
+      sortedOffers = offers.slice().sort((a, b) => b.price - a.price);
+      break;
+    case SortTypes.TOP_RATED_FIRST:
+      sortedOffers = offers.slice().sort((a, b) => b.rating - a.rating);
+      break;
+  }
+
+  return sortedOffers;
+};
+
+export {getRatingWidth, extend, sortOffers};
